roadvision-ai-pro/backend/requirements.txt
# RoadVision AI Pro - Backend Dependencies
# Production-grade AI SaaS Platform

# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# WebSockets
websockets==12.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0

# Computer Vision
opencv-python==4.9.0.80
ultralytics==8.1.2

# ML Dependencies
torch==2.1.2
torchvision==0.16.2
numpy==1.26.3
Pillow==10.2.0

# Payment Integration
stripe==7.10.0

# Utilities
pydantic==2.5.3
python-dotenv==1.0.0
email-validator==2.1.0
python-dateutil==2.8.2

# Monitoring & Logging
prometheus-client==0.19.0
